Fix getnetworkhps overflow bug.

+++ src/rpcmining.cpp
@@ -70,7 +70,7 @@ Value GetNetworkHashPS(int lookup, int height) {
     uint256 workDiff = pb->nChainWork - pb0->nChainWork;
     int64_t timeDiff = maxTime - minTime;
 
-    return (int64_t)(workDiff.getdouble() / timeDiff);
+    return workDiff.getdouble() / timeDiff;
 }
 
 Value getnetworkhashps(const Array& params, bool fHelp)
